database fix and comparison limited to 2 dec points

+++ database.py
@@ -1,4 +1,5 @@
 from sqlmodel import SQLModel, create_engine
+from sqlalchemy import inspect
 from models.cpu import CPU
 from models.gpu import GPU
 from models.motherboard import MotherboardManufacturer, MotherboardChipset, Motherboard
@@ -13,12 +14,18 @@ DATABASE_URL = "sqlite:///./benchmarkinator.db"
 
 engine = create_engine(DATABASE_URL, echo=True)
 
+def check_tables_exist():
+    inspector = inspect(engine)
+    tables = inspector.get_table_names()
+    return "config" in tables and "cpu" in tables and "gpu" in tables and "motherboard" in tables and "ram" in tables and "disk" in tables and "os" in tables and "benchmark" in tables and "benchmarkresult" in tables
+
 def drop_tables():
-    SQLModel.metadata.drop_all(bind=engine)
+    if check_tables_exist():
+        SQLModel.metadata.drop_all(bind=engine)
 
 def init_db():
-    SQLModel.metadata.create_all(bind=engine)
+    if not check_tables_exist():
+        SQLModel.metadata.create_all(bind=engine)
 
-# Drop tables and recreate them
-drop_tables()
+# Initialize the database
 init_db()
